sperate test of RocksDBStateUploader and RocksDBStateDownloader
